requirements.txt
fastapi
uvicorn[standard]
requests
pandas
scikit-learn
joblib
sentence-transformers
faiss-cpu
numpy
python-dotenv
openai
transformers
pydantic
sqlalchemy
alembic
pytest
pytest-asyncio
httpx